pyproject.toml
[tool.poetry]
name = "stock-prediction"
version = "0.0.1"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.9"
amqp = "2.5.1"
aniso8601 = "3.0.2"
appdirs = "1.4.3"
attrs = "19.3.0"
beautifulsoup4 = "4.8.1"
billiard = "3.6.1.0"
bs4 = "0.0.1"
celery = "4.3.0"
certifi = "2019.9.11"
chardet = "3.0.4"
click = "7.0"
cssselect = "1.1.0"
fake-useragent = "0.1.11"
flask-allows = "0.7.1"
flask-restplus = "0.12.1"
graphene = "2.1.3"
graphene-sqlalchemy = "2.1.0"
graphql-core = "2.2.1"
graphql-relay = "0.4.5"
graphql-server-core = "1.1.1"
gunicorn = "19.9.0"
idna = "2.8"
importlib-metadata = "0.23"
itsdangerous = "1.1.0"
jsonschema = "3.1.1"
kombu = "4.6.5"
lxml = "4.9.1"
more-itertools = "7.2.0"
mysqlclient = "1.4.6"
parse = "1.12.1"
promise = "2.2.1"
pyee = "6.0.0"
pyppeteer = "0.0.25"
pyquery = "1.4.1"
pyrsistent = "0.15.4"
python-dateutil = "2.8.0"
pytz = "2019.3"
requests = "2.21.0"
requests-html = "0.10.0"
singledispatch = "3.4.0.3"
six = "1.12.0"
soupsieve = "1.9.4"
tqdm = "4.36.1"
urllib3 = "1.24.3"
vine = "1.3.0"
w3lib = "1.21.0"
websockets = "8.1"
zipp = "0.6.0"
Flask = "1.1.2"
Flask-Cors = "3.0.7"
Flask-GraphQL = "2.0.0"
Flask-JWT-Extended = "3.16.0"
Flask-SQLAlchemy = "2.4.1"
Jinja2 = "2.10.3"
MarkupSafe = "1.1.1"
PyJWT = "1.7.1"
Rx = "1.6.1"
SQLAlchemy = "1.3.16"
Werkzeug = "0.16.0"
pandas = "*"

[tool.poetry.dev-dependencies]
#yapf = "*"
#pigpio = "*"
#smbus2 = "*"
#Tensorflow = "2.1"
#tensorflow-estimator = "^2.1"
#numpy = "*"
#sklearn = "*"
#matplotlib = "^3.2.1"
